requirements.txt
pandas
numpy
scikit-learn
flask
flask-sqlalchemy
pymysql
flask-login
flask-bcrypt
cryptography
gunicorn
alembic
redis
Flask-Caching
Flask-Limiter
pytest
